ES 상품, 게시판 CRUD 작성

+++ src/main/java/com/nhnacademy/marketgg/server/service/impl/DefaultCategoryService.java
@@ -3,18 +3,21 @@ package com.nhnacademy.marketgg.server.service.impl;
 import com.nhnacademy.marketgg.server.dto.request.CategoryCreateRequest;
 import com.nhnacademy.marketgg.server.dto.request.CategoryUpdateRequest;
 import com.nhnacademy.marketgg.server.dto.response.CategoryRetrieveResponse;
+import com.nhnacademy.marketgg.server.elastic.repository.ElasticBoardRepository;
+import com.nhnacademy.marketgg.server.elastic.repository.ElasticProductRepository;
 import com.nhnacademy.marketgg.server.entity.Categorization;
 import com.nhnacademy.marketgg.server.entity.Category;
 import com.nhnacademy.marketgg.server.exception.categorization.CategorizationNotFoundException;
 import com.nhnacademy.marketgg.server.exception.category.CategoryNotFoundException;
 import com.nhnacademy.marketgg.server.repository.categorization.CategorizationRepository;
 import com.nhnacademy.marketgg.server.repository.category.CategoryRepository;
+import com.nhnacademy.marketgg.server.repository.customerservicepost.CustomerServicePostRepository;
+import com.nhnacademy.marketgg.server.repository.product.ProductRepository;
 import com.nhnacademy.marketgg.server.service.CategoryService;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -22,6 +25,10 @@ public class DefaultCategoryService implements CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final CategorizationRepository categorizationRepository;
+    private final ElasticProductRepository elasticProductRepository;
+    private final ElasticBoardRepository elasticBoardRepository;
+    private final ProductRepository productRepository;
+    private final CustomerServicePostRepository customerServicePostRepository;
 
     @Transactional
     @Override
@@ -41,7 +48,8 @@ public class DefaultCategoryService implements CategoryService {
     }
 
     @Override
-    public List<CategoryRetrieveResponse> retrieveCategoriesByCategorization(String categorizationId) {
+    public List<CategoryRetrieveResponse> retrieveCategoriesByCategorization(
+            String categorizationId) {
         return categoryRepository.findByCategorizationCode(categorizationId);
     }
 
@@ -64,10 +72,15 @@ public class DefaultCategoryService implements CategoryService {
     @Transactional
     @Override
     public void deleteCategory(final String id) {
-        Category category = categoryRepository.findById(id)
-                                              .orElseThrow(CategoryNotFoundException::new);
+        Category category = categoryRepository.findById(id).orElseThrow(CategoryNotFoundException::new);
 
+        if (productRepository.existsByCategory(category) ||
+                customerServicePostRepository.existsByCategory(category)) {
+            return;
+        }
         categoryRepository.delete(category);
+        elasticProductRepository.deleteAllByCategoryCode(id);
+        elasticBoardRepository.deleteAllByCategoryCode(id);
     }
 
 }
